fix: ADS Context Menu issues (#39683)
## Description

Fixes two bugs related to ADS Entity Explorer usage of Context Menu

1. To fix the issue with the context menu position, we make sure the
right control has visibility change on hover instead of display. The
menu is anchored to a element for position and when the display of the
element is none, it get attached to the top left corner on 0,0
This is the same solution as the one used in the old entity explorer

https://github.com/appsmithorg/appsmith/blob/release/app/client/src/pages/Editor/Explorer/Entity/index.tsx#L178-L179

2. To fix the issue with the cmd+z not working, we add a timeout for
delete action in the context menu. This is an old issue with menu where
the focus is set to the menu trigger component when the menu closes.
This blocks us from listening to keyboard shortcut events as the element
is not propagating the event up.
This is the same solution as the one used in old entity explorer

https://github.com/appsmithorg/appsmith/blob/release/app/client/src/pages/Editor/Explorer/ContextMenu.tsx#L53-L56

Fixes #39665
Fixes #39641

## Automation

/ok-to-test tags="@tag.IDE"



## Communication
Should the DevRel and Marketing teams inform users about this change?
- [ ] Yes
- [ ] No


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Style**
- Improved hover behavior in list items so that controls reveal
seamlessly without affecting the overall layout.

- **Bug Fixes**
- Refined the widget deletion process to prevent focus issues, ensuring
smoother keyboard shortcut interactions.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->