chore: Utility to clear all sessions in Appsmith (#39693)
## Description
> [!TIP]  
> _Add a TL;DR when the description is longer than 500 words or
extremely technical (helps the content, marketing, and DevRel team)._
>
> _Please also include relevant motivation and context. List any
dependencies that are required for this change. Add links to Notion,
Figma or any other documents that might be relevant to the PR._


Fixes #`Issue Number`  
_or_  
Fixes `Issue URL`
> [!WARNING]  
> _If no issue exists, please create an issue first, and check with the
maintainers if the issue is valid._

## Automation

/ok-to-test tags="@tag.Sanity"

### :mag: Cypress test results
<!-- This is an auto-generated comment: Cypress test results  -->
> [!WARNING]
> Tests have not run on the HEAD
60c365155add5b5f6274ca3a612f1668777580e3 yet
> <hr>Wed, 12 Mar 2025 14:35:59 UTC
<!-- end of auto-generated comment: Cypress test results  -->


## Communication
Should the DevRel and Marketing teams inform users about this change?
- [ ] Yes
- [ ] No


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Enhanced session management now enables the comprehensive removal of
all user sessions—even the current session—when needed.
- This update streamlines session cleanup operations, improves
scalability and system stability, and reduces resource overhead for a
smoother overall experience.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->