chore: remove AWS_ROLE_ARN from deployment workflow and update deploy… (#41039)
… script for node affinity

### Changes Made
- Removed `AWS_ROLE_ARN` from the environment variables in the GitHub
Actions workflow for deploying the Helm chart.
- Cleaned up the `deploy_preview.sh` script by removing commented-out
code related to role assumption.
- Added a new `dp-node-affinity-values.yaml` file to define node
affinity and tolerations for deployment.

These changes streamline the deployment process and enhance the
configuration for node scheduling.

## Description
> [!TIP]  
> _Add a TL;DR when the description is longer than 500 words or
extremely technical (helps the content, marketing, and DevRel team)._
>
> _Please also include relevant motivation and context. List any
dependencies that are required for this change. Add links to Notion,
Figma or any other documents that might be relevant to the PR._


Fixes #`Issue Number`  
_or_  
Fixes `Issue URL`
> [!WARNING]  
> _If no issue exists, please create an issue first, and check with the
maintainers if the issue is valid._

## Automation

/ok-to-test tags=""

### :mag: Cypress test results
<!-- This is an auto-generated comment: Cypress test results  -->
> [!CAUTION]  
> If you modify the content in this section, you are likely to disrupt
the CI result for your PR.

<!-- end of auto-generated comment: Cypress test results  -->


## Communication
Should the DevRel and Marketing teams inform users about this change?
- [ ] Yes
- [ ] No