feat: allow specifying deployment in helm chart without autoscaling (#41397)
## Description

Allows specifying that the deployment of Appsmith should be done via the
Deployment controller without enabling autoscaling. This is useful in
situations where you want to be able to set a number of replicas without
using HPAs or if you want the flexibility to set replica count to 0 to
temporarily pause the instance.

Fixes #`Issue Number`  
_or_  
Fixes `Issue URL`
> [!WARNING]  
> _If no issue exists, please create an issue first, and check with the
maintainers if the issue is valid._

## Automation

/ok-to-test tags=""

### :mag: Cypress test results
<!-- This is an auto-generated comment: Cypress test results  -->
> [!WARNING]
> Tests have not run on the HEAD
35fba39e7724faf3710ea1717dcbec0a5b91553e yet
> <hr>Wed, 19 Nov 2025 03:42:56 UTC
<!-- end of auto-generated comment: Cypress test results  -->


## Communication
Should the DevRel and Marketing teams inform users about this change?
- [ ] Yes
- [x] No


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

* **New Features**
* Choice of workload type (Deployment or StatefulSet) and replicas
behavior when using Deployment
* Persistent storage now can be provisioned for Deployment workloads as
well as StatefulSets

* **Documentation**
* Added guidance on workload kind, replicas usage, and interaction with
autoscaling

* **Tests**
* Added tests covering persistence and workload-kind rendering scenarios

* **Chores**
  * Chart version bumped to 3.6.6
<!-- end of auto-generated comment: release notes by coderabbit.ai -->